+++ lib/ios/RNNEventEmitter.m
@@ -11,27 +11,31 @@ RCT_EXPORT_MODULE();
 static NSString* const AppLaunched				= @"RNN.AppLaunched";
 static NSString* const CommandCompleted			= @"RNN.CommandCompleted";
 static NSString* const BottomTabSelected		= @"RNN.BottomTabSelected";
+static NSString* const BottomTabLongPressed     = @"RNN.BottomTabLongPressed";
 static NSString* const ComponentDidAppear		= @"RNN.ComponentDidAppear";
 static NSString* const ComponentDidDisappear	= @"RNN.ComponentDidDisappear";
 static NSString* const NavigationButtonPressed	= @"RNN.NavigationButtonPressed";
 static NSString* const ModalDismissed	        = @"RNN.ModalDismissed";
+static NSString* const ModalAttemptedToDismiss  = @"RNN.ModalAttemptedToDismiss";
 static NSString* const SearchBarUpdated 		= @"RNN.SearchBarUpdated";
 static NSString* const SearchBarCancelPressed 	= @"RNN.SearchBarCancelPressed";
 static NSString* const PreviewCompleted         = @"RNN.PreviewCompleted";
 static NSString* const ScreenPopped             = @"RNN.ScreenPopped";
 
-- (NSArray<NSString *> *)supportedEvents {
-    return @[AppLaunched,
-             CommandCompleted,
-             BottomTabSelected,
-             ComponentDidAppear,
-             ComponentDidDisappear,
-             NavigationButtonPressed,
-             ModalDismissed,
-             SearchBarUpdated,
-             SearchBarCancelPressed,
-             PreviewCompleted,
-             ScreenPopped];
+-(NSArray<NSString *> *)supportedEvents {
+	return @[AppLaunched,
+			 CommandCompleted,
+			 BottomTabSelected,
+       BottomTabLongPressed,
+			 ComponentDidAppear,
+			 ComponentDidDisappear,
+			 NavigationButtonPressed,
+			 ModalDismissed,
+			 SearchBarUpdated,
+			 SearchBarCancelPressed,
+			 PreviewCompleted,
+       ScreenPopped,
+       ModalAttemptedToDismiss];
 }
 
 # pragma mark public
@@ -44,17 +48,19 @@ static NSString* const ScreenPopped             = @"RNN.ScreenPopped";
     }
 }
 
-- (void)sendComponentDidAppear:(NSString *)componentId componentName:(NSString *)componentName {
+- (void)sendComponentDidAppear:(NSString *)componentId componentName:(NSString *)componentName componentType:(NSString *)componentType {
     [self send:ComponentDidAppear body:@{
         @"componentId":componentId,
-        @"componentName": componentName
+        @"componentName": componentName,
+        @"componentType": componentType
     }];
 }
 
-- (void)sendComponentDidDisappear:(NSString *)componentId componentName:(NSString *)componentName{
+- (void)sendComponentDidDisappear:(NSString *)componentId componentName:(NSString *)componentName componentType:(NSString *)componentType {
     [self send:ComponentDidDisappear body:@{
         @"componentId":componentId,
-        @"componentName": componentName
+        @"componentName": componentName,
+        @"componentType": componentType
     }];
 }
 
@@ -72,13 +78,19 @@ static NSString* const ScreenPopped             = @"RNN.ScreenPopped";
     }];
 }
 
-- (void)sendOnNavigationCommandCompletion:(NSString *)commandName commandId:(NSString *)commandId params:(NSDictionary*)params {
-    [self send:CommandCompleted body:@{
-        @"commandId":commandId,
-        @"commandName":commandName,
-        @"params": params,
-        @"completionTime": [RNNUtils getCurrentTimestamp]
-    }];
+- (void)sendBottomTabLongPressed:(NSNumber *)selectedTabIndex {
+    [self send:BottomTabLongPressed body:@{
+                                        				  @"selectedTabIndex": selectedTabIndex
+                                        				  }];
+}
+
+-(void)sendOnNavigationCommandCompletion:(NSString *)commandName commandId:(NSString *)commandId params:(NSDictionary*)params {
+	[self send:CommandCompleted body:@{
+									   @"commandId":commandId,
+									   @"commandName":commandName,
+									   @"params": params,
+									   @"completionTime": [RNNUtils getCurrentTimestamp]
+									   }];
 }
 
 - (void)sendOnSearchBarUpdated:(NSString *)componentId
@@ -108,6 +120,12 @@ static NSString* const ScreenPopped             = @"RNN.ScreenPopped";
     [self send:ModalDismissed body:@{
         @"componentId": componentId,
         @"modalsDismissed": modalsDismissed
+    }];
+}
+
+- (void)sendModalAttemptedToDismissEvent:(NSString *)componentId {
+    [self send:ModalAttemptedToDismiss body:@{
+        @"componentId": componentId,
     }];
 }