+++ lib/ios/RNNReactView.m
@@ -3,69 +3,63 @@
 #import <React/RCTUIManager.h>
 
 @implementation RNNReactView {
-	BOOL _fillParent;
+    BOOL _isAppeared;
 }
 
-- (instancetype)initWithBridge:(RCTBridge *)bridge moduleName:(NSString *)moduleName initialProperties:(NSDictionary *)initialProperties availableSize:(CGSize)availableSize reactViewReadyBlock:(RNNReactViewReadyCompletionBlock)reactViewReadyBlock {
+- (instancetype)initWithBridge:(RCTBridge *)bridge moduleName:(NSString *)moduleName initialProperties:(NSDictionary *)initialProperties eventEmitter:(RNNEventEmitter *)eventEmitter reactViewReadyBlock:(RNNReactViewReadyCompletionBlock)reactViewReadyBlock {
 	self = [super initWithBridge:bridge moduleName:moduleName initialProperties:initialProperties];
 	[[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(contentDidAppear:) name:RCTContentDidAppearNotification object:nil];
 	 _reactViewReadyBlock = reactViewReadyBlock;
-    [bridge.uiManager setAvailableSize:UIScreen.mainScreen.bounds.size forRootView:self];
+    _eventEmitter = eventEmitter;
     
 	return self;
 }
 
+- (void)layoutSubviews {
+    [super layoutSubviews];
+    #ifdef DEBUG
+        [RCTHelpers removeYellowBox:self];
+    #endif
+}
+
 - (void)contentDidAppear:(NSNotification *)notification {
-#ifdef DEBUG
-	if ([((RNNReactView *)notification.object).moduleName isEqualToString:self.moduleName]) {
-		[RCTHelpers removeYellowBox:self];
-	}
-#endif
-	
 	RNNReactView* appearedView = notification.object;
-	
-	 if (_reactViewReadyBlock && [appearedView.appProperties[@"componentId"] isEqual:self.componentId]) {
-	 	_reactViewReadyBlock();
-		 _reactViewReadyBlock = nil;
-		 [[NSNotificationCenter defaultCenter] removeObserver:self];
+	 if ([appearedView.appProperties[@"componentId"] isEqual:self.componentId]) {
+         [self reactViewReady];
 	 }
+}
+
+- (void)reactViewReady {
+    if (_reactViewReadyBlock) {
+        _reactViewReadyBlock();
+        _reactViewReadyBlock = nil;
+    }
+    [[NSNotificationCenter defaultCenter] removeObserver:self];
+}
+
+
+- (void)componentDidAppear {
+    if (!_isAppeared) {
+        [_eventEmitter sendComponentDidAppear:self.componentId componentName:self.moduleName componentType:self.componentType];
+    }
+    
+    _isAppeared = YES;
+}
+
+- (void)componentDidDisappear {
+    if (_isAppeared) {
+        [_eventEmitter sendComponentDidDisappear:self.componentId componentName:self.moduleName componentType:self.componentType];
+    }
+    
+    _isAppeared = NO;
 }
 
 - (NSString *)componentId {
 	return self.appProperties[@"componentId"];
 }
 
-- (void)setRootViewDidChangeIntrinsicSize:(void (^)(CGSize))rootViewDidChangeIntrinsicSize {
-		_rootViewDidChangeIntrinsicSize = rootViewDidChangeIntrinsicSize;
-		self.delegate = self;
-}
-
-- (void)rootViewDidChangeIntrinsicSize:(RCTRootView *)rootView {
-	if (_rootViewDidChangeIntrinsicSize) {
-		_rootViewDidChangeIntrinsicSize(rootView.intrinsicContentSize);
-	}
-}
-
-- (CGSize)intrinsicContentSize {
-	if (_fillParent) {
-		return UILayoutFittingExpandedSize;
-	} else {
-		return [super intrinsicContentSize];
-	}
-}
-
-- (void)setAlignment:(NSString *)alignment inFrame:(CGRect)frame {
-	if ([alignment isEqualToString:@"fill"]) {
-		_fillParent = YES;
-		self.translatesAutoresizingMaskIntoConstraints = NO;
-		self.sizeFlexibility = RCTRootViewSizeFlexibilityNone;
-	} else {
-		self.sizeFlexibility = RCTRootViewSizeFlexibilityWidthAndHeight;
-		__weak RNNReactView *weakSelf = self;
-		[self setRootViewDidChangeIntrinsicSize:^(CGSize intrinsicSize) {
-			[weakSelf setFrame:CGRectMake(0, 0, intrinsicSize.width, intrinsicSize.height)];
-		}];
-	}
+- (NSString *)componentType {
+    return ComponentTypeScreen;
 }
 
 @end