+++ lib/ios/UIViewController+LayoutProtocol.m
@@ -80,11 +80,18 @@
 }
 
 - (UIViewController *)getCurrentChild {
+    for (UIViewController* childViewController in self.childViewControllers.reverseObjectEnumerator.allObjects) {
+        if (childViewController.layoutInfo) {
+            return childViewController;
+        }
+    }
+    
     return nil;
 }
 
 - (UIViewController *)presentedComponentViewController {
-    return self.getCurrentChild ? self.getCurrentChild.presentedComponentViewController : self;
+    UIViewController* currentChild = self.getCurrentChild;
+    return currentChild ? currentChild.presentedComponentViewController : self;
 }
 
 - (UIViewController *)topMostViewController {
@@ -114,7 +121,17 @@
 
 - (void)onChildWillAppear {
 	[self.presenter applyOptions:self.resolveOptions];
-	[((UISplitViewController *)self.parentViewController) onChildWillAppear];
+	[self.parentViewController onChildWillAppear];
+}
+
+- (void)componentDidAppear {
+    [self.presenter componentDidAppear];
+    [self.parentViewController componentDidAppear];
+}
+
+- (void)componentDidDisappear {
+    [self.presenter componentDidDisappear];
+    [self.parentViewController componentDidDisappear];
 }
 
 - (void)willMoveToParentViewController:(UIViewController *)parent {