+++ lib/src/events/ComponentEventsObserver.test.tsx
@@ -18,6 +18,7 @@ describe('ComponentEventsObserver', () => {
   const searchBarCancelPressedFn = jest.fn();
   const previewCompletedFn = jest.fn();
   const modalDismissedFn = jest.fn();
+  const modalAttemptedToDismissFn = jest.fn();
   const screenPoppedFn = jest.fn();
   let subscription: EventSubscription;
   let uut: ComponentEventsObserver;
@@ -55,6 +56,10 @@ describe('ComponentEventsObserver', () => {
 
     modalDismissed(event: any) {
       modalDismissedFn(event);
+    }
+
+    modalAttemptedToDismiss(event: any) {
+      modalAttemptedToDismissFn(event);
     }
 
     searchBarUpdated(event: any) {
@@ -108,6 +113,10 @@ describe('ComponentEventsObserver', () => {
       modalDismissedFn(event);
     }
 
+    modalAttemptedToDismiss(event: any) {
+      modalAttemptedToDismissFn(event);
+    }
+
     searchBarUpdated(event: any) {
       searchBarUpdatedFn(event);
     }
@@ -148,21 +157,21 @@ describe('ComponentEventsObserver', () => {
     expect(tree.toJSON()).toBeDefined();
     expect(didAppearFn).not.toHaveBeenCalled();
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
   });
 
   it(`bindComponent should use optional componentId if component has a componentId in props`, () => {
-    const tree = renderer.create(<UnboundScreen  componentId={'doNotUseThisId'} />);
+    const tree = renderer.create(<UnboundScreen componentId={'doNotUseThisId'} />);
     uut.bindComponent(tree.getInstance() as any, 'myCompId')
 
     expect(tree.toJSON()).toBeDefined();
-    
-    uut.notifyComponentDidAppear({ componentId: 'dontUseThisId', componentName: 'doesnt matter' });
-    expect(didAppearFn).not.toHaveBeenCalled();
-    
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'dontUseThisId', componentName: 'doesnt matter', componentType: 'Component' });
+    expect(didAppearFn).not.toHaveBeenCalled();
+
+
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
   });
 
@@ -174,10 +183,10 @@ describe('ComponentEventsObserver', () => {
     expect(didDisappearFn).not.toHaveBeenCalled();
     expect(willUnmountFn).not.toHaveBeenCalled();
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
 
-    uut.notifyComponentDidDisappear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidDisappear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didDisappearFn).toHaveBeenCalledTimes(1);
 
     uut.notifyNavigationButtonPressed({ componentId: 'myCompId', buttonId: 'myButtonId' });
@@ -187,6 +196,10 @@ describe('ComponentEventsObserver', () => {
     uut.notifyModalDismissed({ componentId: 'myCompId', modalsDismissed: 1 });
     expect(modalDismissedFn).toHaveBeenCalledTimes(1);
     expect(modalDismissedFn).toHaveBeenLastCalledWith({ componentId: 'myCompId', modalsDismissed: 1 })
+
+    uut.notifyModalAttemptedToDismiss({ componentId: 'myCompId' });
+    expect(modalAttemptedToDismissFn).toHaveBeenCalledTimes(1);
+    expect(modalAttemptedToDismissFn).toHaveBeenLastCalledWith({ componentId: 'myCompId' })
 
     uut.notifySearchBarUpdated({ componentId: 'myCompId', text: 'theText', isFocused: true });
     expect(searchBarUpdatedFn).toHaveBeenCalledTimes(1);
@@ -211,6 +224,7 @@ describe('ComponentEventsObserver', () => {
   it(`componentDidAppear should receive component props from store`, () => {
     const event = {
       componentId: 'myCompId',
+      componentType: 'Component',
       passProps: {
         propA: 'propA'
       },
@@ -220,14 +234,14 @@ describe('ComponentEventsObserver', () => {
     mockStore.updateProps(event.componentId, event.passProps)
     expect(didAppearFn).not.toHaveBeenCalled();
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
     expect(didAppearFn).toHaveBeenCalledWith(event);
   });
 
   it(`doesnt call other componentIds`, () => {
     renderer.create(<BoundScreen componentId={'myCompId'} />);
-    uut.notifyComponentDidAppear({ componentId: 'other', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'other', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).not.toHaveBeenCalled();
   });
 
@@ -235,18 +249,18 @@ describe('ComponentEventsObserver', () => {
     const tree = renderer.create(<SimpleScreen componentId={'myCompId'} />);
     expect((tree.getInstance() as any).componentDidAppear).toBeUndefined();
     uut.bindComponent(tree.getInstance() as any);
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
   });
 
   it(`returns unregister fn`, () => {
     renderer.create(<BoundScreen componentId={'123'} />);
 
-    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
 
     subscription.remove();
 
-    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).toHaveBeenCalledTimes(1);
   });
 
@@ -256,7 +270,7 @@ describe('ComponentEventsObserver', () => {
 
     uut.unmounted('123');
 
-    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: '123', componentName: 'doesnt matter', componentType: 'Component' });
     expect(didAppearFn).not.toHaveBeenCalled();
   });
 
@@ -272,20 +286,20 @@ describe('ComponentEventsObserver', () => {
     const result2 = uut.bindComponent(instance2);
     expect(result1).not.toEqual(result2);
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
 
     expect(instance1.componentDidAppear).toHaveBeenCalledTimes(1);
     expect(instance2.componentDidAppear).toHaveBeenCalledTimes(1);
 
     result2.remove();
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(instance1.componentDidAppear).toHaveBeenCalledTimes(2);
     expect(instance2.componentDidAppear).toHaveBeenCalledTimes(1);
 
     result1.remove();
 
-    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter' });
+    uut.notifyComponentDidAppear({ componentId: 'myCompId', componentName: 'doesnt matter', componentType: 'Component' });
     expect(instance1.componentDidAppear).toHaveBeenCalledTimes(2);
     expect(instance2.componentDidAppear).toHaveBeenCalledTimes(1);
   });