+++ lib/src/events/EventsRegistry.ts
@@ -12,9 +12,10 @@ import {
   SearchBarCancelPressedEvent,
   PreviewCompletedEvent,
   ModalDismissedEvent,
-  ScreenPoppedEvent
+  ScreenPoppedEvent,
+  ModalAttemptedToDismissEvent
 } from '../interfaces/ComponentEvents';
-import { CommandCompletedEvent, BottomTabSelectedEvent } from '../interfaces/Events';
+import { CommandCompletedEvent, BottomTabSelectedEvent, BottomTabLongPressedEvent } from '../interfaces/Events';
 
 export class EventsRegistry {
   constructor(private nativeEventsReceiver: NativeEventsReceiver, private commandsObserver: CommandsObserver, private componentEventsObserver: ComponentEventsObserver) { }
@@ -39,12 +40,20 @@ export class EventsRegistry {
     return this.nativeEventsReceiver.registerBottomTabSelectedListener(callback);
   }
 
+  public registerBottomTabLongPressedListener(callback: (event: BottomTabLongPressedEvent) => void): EmitterSubscription {
+    return this.nativeEventsReceiver.registerBottomTabLongPressedListener(callback);
+  }
+
   public registerNavigationButtonPressedListener(callback: (event: NavigationButtonPressedEvent) => void): EmitterSubscription {
     return this.nativeEventsReceiver.registerNavigationButtonPressedListener(callback);
   }
 
   public registerModalDismissedListener(callback: (event: ModalDismissedEvent) => void): EmitterSubscription {
     return this.nativeEventsReceiver.registerModalDismissedListener(callback);
+  }
+
+  public registerModalAttemptedToDismissListener(callback: (event: ModalAttemptedToDismissEvent) => void): EmitterSubscription {
+    return this.nativeEventsReceiver.registerModalAttemptedToDismissListener(callback);
   }
 
   public registerSearchBarUpdatedListener(callback: (event: SearchBarUpdatedEvent) => void): EmitterSubscription {