+++ lib/src/interfaces/ComponentEvents.ts
@@ -1,3 +1,5 @@
+export type ComponentType = 'Component' | 'TopBarTitle' | 'TopBarBackground' | 'TopBarButton';
+
 export interface ComponentEvent {
   componentId: string;
 }
@@ -5,10 +7,12 @@ export interface ComponentEvent {
 export interface ComponentDidAppearEvent extends ComponentEvent {
   componentName: string;
   passProps?: object;
+  componentType: ComponentType;
 }
 
 export interface ComponentDidDisappearEvent extends ComponentEvent {
   componentName: string;
+  componentType: ComponentType;
 }
 
 export interface NavigationButtonPressedEvent extends ComponentEvent {
@@ -18,6 +22,10 @@ export interface NavigationButtonPressedEvent extends ComponentEvent {
 export interface ModalDismissedEvent extends ComponentEvent {
   componentId: string;
   modalsDismissed: number;
+}
+
+export interface ModalAttemptedToDismissEvent extends ComponentEvent {
+  componentId: string;
 }
 
 export interface SearchBarUpdatedEvent extends ComponentEvent {